QBRI123-76 add new class HeaderFragment and modified tests
