+++ src/test/java/tests/SearchProductTest.java
@@ -4,7 +4,7 @@ import base.AbstractBaseTest;
 import org.assertj.core.api.SoftAssertions;
 import org.testng.annotations.Test;
 import pages.BrowseProductsPage;
-import pages.HomePage;
+import pages.HeaderFragment;
 
 import java.util.List;
 
@@ -21,11 +21,11 @@ public class SearchProductTest extends AbstractBaseTest {
 
     @Test(description = "Product search by name")
     public void searchProductByName() {
-        HomePage homePage = new HomePage(driver);
+        HeaderFragment headerFragment = new HeaderFragment(driver);
         BrowseProductsPage browseProductsPage = new BrowseProductsPage(driver);
         SoftAssertions softAssert = new SoftAssertions();
 
-        homePage.fillBaseSearchField(nameOfSearchProduct);
+        headerFragment.fillBaseSearchField(nameOfSearchProduct);
 
         listOfNames = browseProductsPage.getTitlesNames();
         for (String name : listOfNames) {
@@ -36,11 +36,11 @@ public class SearchProductTest extends AbstractBaseTest {
 
     @Test(description = "Using filter in search")
     public void searchesProductByName() {
-        HomePage homePage = new HomePage(driver);
+        HeaderFragment headerFragment = new HeaderFragment(driver);
         BrowseProductsPage browseProductsPage = new BrowseProductsPage(driver);
         SoftAssertions soft = new SoftAssertions();
 
-        homePage.fillBaseSearchField(nameOfSearchProducts);
+        headerFragment.fillBaseSearchField(nameOfSearchProducts);
         browseProductsPage.selectFilterOption(toddler);
         browseProductsPage.waitLoading();
         genderFilterText = browseProductsPage.getGenderField(toddler);
